meshgraphnets/requirements.txt
# Modified from original to get rid of Sonnet
tensorflow-gpu
matplotlib
absl-py
numpy
